pvalue chart filter / need to add listener on slider
